Change type of SW_VERSION from uint8_t to uint16_t
This change makes sure that MAJOR_VERSION and MINOR_VERSION can get
bigger than 15 and differentiates this version from the original one.

+++ src/wceeprom.h
@@ -95,7 +95,7 @@ typedef struct WcEepromData {
      * @see MAJOR_VERSION()
      * @see MINOR_VERSION()
      */
-    uint8_t swVersion;
+    uint16_t swVersion;
 
     /**
      * @brief Size of this struct
